chore: refactor live conversation
