+++ src/core/client/stream/tabs/Live/LiveConversation/LiveCreateCommentReplyMutation.tsx
@@ -1,4 +1,3 @@
-import { pick } from "lodash";
 import { Environment, graphql } from "react-relay";
 import { RecordSourceSelectorProxy } from "relay-runtime";
 
@@ -11,7 +10,7 @@ import {
   lookup,
   MutationInput,
 } from "coral-framework/lib/relay";
-import { GQLComment, GQLStory, GQLUSER_ROLE } from "coral-framework/schema";
+import { GQLStory, GQLUSER_ROLE } from "coral-framework/schema";
 import { CreateCommentReplyEvent } from "coral-stream/events";
 
 import { LiveCreateCommentReplyMutation as MutationTypes } from "coral-stream/__generated__/LiveCreateCommentReplyMutation.graphql";
@@ -86,6 +85,7 @@ function sharedUpdater(
     return;
   }
   const node = commentEdge.getLinkedRecord("node")!;
+  commentEdge.setValue(node.getValue("createdAt"), "cursor");
 
   insertCommentToStory(store, input.storyID, node, {
     liveInsertion: false,
@@ -102,7 +102,6 @@ async function commit(
   input: LiveCreateCommentReplyInput,
   { uuidGenerator, relayEnvironment, eventEmitter }: CoralContext
 ) {
-  const parentComment = lookup<GQLComment>(environment, input.parentID)!;
   const viewer = getViewer<LiveCreateCommentReplyMutation_viewer>(environment)!;
   const currentDate = new Date().toISOString();
   const id = uuidGenerator();
@@ -164,12 +163,7 @@ async function commit(
                   media: null,
                 },
                 parent: {
-                  createdAt: parentComment.createdAt,
-                  id: parentComment.id,
-                  author: parentComment.author
-                    ? pick(parentComment.author, "username", "id")
-                    : null,
-                  body: parentComment.body,
+                  id: input.parentID,
                 },
                 editing: {
                   editableUntil: new Date(Date.now() + 10000).toISOString(),
