+++ src/core/client/stream/tabs/Live/LiveConversation/LiveReplyContainer.tsx
@@ -14,9 +14,9 @@ import { LiveReplyContainer_settings } from "coral-stream/__generated__/LiveRepl
 import { LiveReplyContainer_story } from "coral-stream/__generated__/LiveReplyContainer_story.graphql";
 import { LiveReplyContainer_viewer } from "coral-stream/__generated__/LiveReplyContainer_viewer.graphql";
 
-import InView from "../../InView";
-import LiveCommentActionsContainer from "../../LiveComment/LiveCommentActionsContainer";
-import LiveCommentBodyContainer from "../../LiveComment/LiveCommentBodyContainer";
+import InView from "../InView";
+import LiveCommentActionsContainer from "../LiveComment/LiveCommentActionsContainer";
+import LiveCommentBodyContainer from "../LiveComment/LiveCommentBodyContainer";
 
 import styles from "./LiveReplyContainer.css";
